requirements.txt
mysql_connector_repackaged==0.3.1
numpy==2.2.4
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.2.3
pyserial==3.5
pyttsx3==2.98
streamlit==1.43.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
twilio==9.5.1
